Improved design to deal with different distributions
Before, the code checked for a specific distribution
type and made a downcast to the derived type. This
is IMHO an anti-patters due to numeral reasons:
 1. If a new type is added, it is difficult and error prone to find
    all places in the code that check and downcast
 2. If the type enum is incorrectly set, it also won't work

This change uses the visitor pattern. It improves the current
design as the compiler now checks everything. If a new
distribution is added, all places that have to be adapted
also won't compile anymore.

Closes #135