Small model adaptation
In the current implementation, in secir_get_derivates
it was divided by N0 which always only
referred to the start values and additionally also contained
the dead people at the start of the beginning.
It is now corrected that N does not contain dead people and it always
refers to the current number of living people.

The comparison data file had to be adapted.