916-refactor-epidata-to-support-ingestion-into-the-LOKI-database (#1024)
Co-authored-by: Mariama Jaiteh <mariama.jaiteh@helmholtz-hzi.de>

This PR is the refactoring of the memilio-epidata package to support data ingestion using python objects rather than files. No change has been made to preexisting functionalities. A CLI args was added to choose the option of creating objects. This PR allows integration with external packages of the LOKI project.